backend/requirements.txt
# Requires Python >= 3.10 (see setup-local.sh). Use e.g. brew install python@3.12
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
argcomplete==3.6.2
asyncpg==0.30.0
attrs==25.3.0
beautifulsoup4==4.13.5
bleach==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
# 8.3.x requires Python >=3.10; 8.1.8 works on 3.9+ (use Python 3.10+ for this project when possible)
click==8.1.8
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.116.2
frozenlist==1.7.0
future==1.0.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jiter==0.11.0
jq==1.10.0
jsonpatch==1.33
jsonpointer==3.0.0
kaggle==1.7.4.5
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-ollama>=0.2.0
langchain-postgres==0.0.15
langchain-text-splitters==0.3.11
langsmith==0.4.28
marshmallow==3.26.1
multidict==6.6.4
mypy_extensions==1.1.0
numpy==2.3.3
openai==1.108.0
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pgvector==0.3.6
pillow==11.3.0
propcache==0.3.2
protobuf==6.32.1
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg-pool==3.2.6
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyMuPDF==1.26.4
pytesseract==0.3.13
pdf2image==1.17.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
retriever==3.1.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
starlette==0.48.0
tenacity==9.1.2
text-unidecode==1.3
tiktoken==0.11.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
webencodings==0.5.1
xlrd==2.0.2
yarl==1.20.1
zstandard==0.25.0
